requirements.txt
firecrawl-py>=0.0.8
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
asyncio-throttle>=1.0.2
tiktoken>=0.7.0
aiohttp>=3.9.0
